Start refactor of global state related things to src/state
