+++ src/api/actions.tsx
@@ -7,13 +7,10 @@ import type {
   Event,
   AcknowledgementBody,
   Acknowledgement,
-  AutoUpdateMethod,
-  User,
-  Token,
 } from "../api/types.d";
 
 import api from "../api";
-import { InitialStateType, ActionsType } from "../contexts";
+import type { InitialStateType, ActionsType } from "../state/contexts";
 import {
   createFilter as createFilterAction,
   deleteFilter as deleteFilterAction,
@@ -25,16 +22,7 @@ import { IncidentsFilter } from "../components/incidenttable/FilteredIncidentTab
 import { Tag } from "../components/tagselector";
 import { IncidentsStateType, useIncidentsContext } from "../components/incidentsprovider";
 
-import {
-  ApiState,
-  setAutoUpdateMethod,
-  setHasConnectionProblems,
-  unsetHasConnectionProblems,
-} from "../reducers/apistate";
-
-import { UserStateType, loginUser, loginTokenUser, logoutUser } from "../reducers/user";
-
-import { AppContext } from "../contexts";
+import { AppContext } from "../state/contexts";
 
 type Dispatch = React.Dispatch<ActionsType>;
 
@@ -163,68 +151,6 @@ export const useIncidents = (): [IncidentsStateType, UseIncidentsActionType] => 
       closeIncident: closeIncidentCallback,
       reopenIncident: reopenIncidentCallback,
       acknowledgeIncident: acknowledgeIncidentCallback,
-    },
-  ];
-};
-
-/*
- * TODO: Move somewhere else
- */
-export type UseApiStateActionType = {
-  setAutoUpdateMethod: (method: AutoUpdateMethod) => void;
-  setHasConnectionProblems: () => void;
-  unsetHasConnectionProblems: () => void;
-};
-
-export const useApiState = (): [ApiState, UseApiStateActionType] => {
-  const {
-    state: { apiState },
-    dispatch,
-  } = useContext(AppContext);
-
-  const setAutoUpdateMethodCallback = useCallback((method: AutoUpdateMethod) => dispatch(setAutoUpdateMethod(method)), [
-    dispatch,
-  ]);
-  const setHasConnectionProblemsCallback = useCallback(() => dispatch(setHasConnectionProblems()), [dispatch]);
-  const unsetHasConnectionProblemsCallback = useCallback(() => dispatch(unsetHasConnectionProblems()), [dispatch]);
-
-  return [
-    apiState,
-    {
-      setAutoUpdateMethod: setAutoUpdateMethodCallback,
-      setHasConnectionProblems: setHasConnectionProblemsCallback,
-      unsetHasConnectionProblems: unsetHasConnectionProblemsCallback,
-    },
-  ];
-};
-
-/*
- * User
- */
-export type UseUserActionType = {
-  login: (user: User) => void;
-  logout: () => void;
-  loginToken: (user: User, token: Token) => void;
-};
-
-export const useUser = (): [UserStateType, UseUserActionType] => {
-  const {
-    state: { user },
-    dispatch,
-  } = useContext(AppContext);
-
-  const loginCallback = useCallback((user: User) => dispatch(loginUser(user)), [dispatch]);
-  const logoutCallback = useCallback(() => dispatch(logoutUser()), [dispatch]);
-  const loginTokenCallback = useCallback((user: User, token: Token) => dispatch(loginTokenUser({ user, token })), [
-    dispatch,
-  ]);
-
-  return [
-    user,
-    {
-      login: loginCallback,
-      logout: logoutCallback,
-      loginToken: loginTokenCallback,
     },
   ];
 };
