+++ src/components/header/Header.tsx
@@ -18,7 +18,7 @@ import Skeleton from "@material-ui/lab/Skeleton";
 import auth from "../../auth";
 
 // Contexts/Hooks
-import { useApiState, useUser } from "../../api/actions";
+import { useApiState, useUser } from "../../state/hooks";
 
 // Components
 import Menu from "../../components/menu";
