+++ src/index.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom";
 import "./index.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
-import { AppProvider } from "./contexts";
+import { AppProvider } from "./state/contexts";
 // TODO: Remove use of store provider
 import { StoreProvider } from "./store";
 import { BrowserRouter } from "react-router-dom";
