+++ src/state/contexts.tsx
@@ -1,9 +1,9 @@
 import React, { createContext, useReducer } from "react";
 
-import type { Filter } from "./api/types.d";
-import { filterReducer, FilterActions } from "./reducers/filter";
-import { initialUserState, userReducer, UserActions, UserStateType } from "./reducers/user";
-import { initialApiState, apiStateReducer, ApiStateActions, ApiState } from "./reducers/apistate";
+import type { Filter } from "../api/types.d";
+import { filterReducer, FilterActions } from "../reducers/filter";
+import { initialUserState, userReducer, UserActions, UserStateType } from "../reducers/user";
+import { initialApiState, apiStateReducer, ApiStateActions, ApiState } from "../reducers/apistate";
 
 export type InitialStateType = {
   // List of all filters that the currently
